+++ src/app/dialogs/components/dialog.base.ts
