+++ src/app/dialogs/components/dialog.main.component.ts
