+++ src/app/dialogs/components/dialogs/about.scss
