+++ src/app/dialogs/components/dialogs/about.ts
