+++ src/app/dialogs/components/dialogs/custom.type.ts
