+++ src/app/dialogs/components/dialogs/show.table.properties.ts
