+++ src/app/sql/beans/enumeration.ts
